tests/wpt/tests/html/canvas/element/shadows/2d.shadow.attributes.shadowBlur.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.shadow.attributes.shadowBlur.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.shadow.attributes.shadowBlur.invalid</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

  ctx.shadowBlur = 1;
  ctx.shadowBlur = -2;
  _assertSame(ctx.shadowBlur, 1, "ctx.shadowBlur", "1");

  ctx.shadowBlur = 1;
  ctx.shadowBlur = Infinity;
  _assertSame(ctx.shadowBlur, 1, "ctx.shadowBlur", "1");

  ctx.shadowBlur = 1;
  ctx.shadowBlur = -Infinity;
  _assertSame(ctx.shadowBlur, 1, "ctx.shadowBlur", "1");

  ctx.shadowBlur = 1;
  ctx.shadowBlur = NaN;
  _assertSame(ctx.shadowBlur, 1, "ctx.shadowBlur", "1");

  ctx.shadowBlur = 1;
  ctx.shadowBlur = 'string';
  _assertSame(ctx.shadowBlur, 1, "ctx.shadowBlur", "1");

  ctx.shadowBlur = 1;
  ctx.shadowBlur = true;
  _assertSame(ctx.shadowBlur, 1, "ctx.shadowBlur", "1");

  ctx.shadowBlur = 1;
  ctx.shadowBlur = false;
  _assertSame(ctx.shadowBlur, 0, "ctx.shadowBlur", "0");

});
</script>
